requirements.txt
# Base requirements for Xorb platform
# Core web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.1.0

# HTTP client and async support
httpx==0.27.0
asyncio-mqtt==0.14.0

# Database drivers
asyncpg==0.29.0
redis==5.0.1

# Workflow orchestration
temporalio==1.5.1

# Monitoring and metrics
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.0.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
nvdlib==0.7.1
